modified

+++ main.py
@@ -5,7 +5,7 @@ from src.bigdata1.api import get_data
 
 if __name__ == "__main__":
 	parse = argparse.ArgumentParser(description="Process parameters")
-	parse = add_argument('--page_size', type = float)
+	parse = add_argument('--page_size', type = int)
 
 	args = parser.parse_args()
 	get_data(arg.page_size)
